Add script to export the complete apps tree
Each of the individual components is a git-submodule, and git-archive
only works with a single repository. To generate the full tree from the
master tree, and all the included submodules, we need to explicitly
construct complete list of files contained within the hierarchy.

This script generates a tarball from the complete repository.