Create per-subproject project file symlinks
Qmake depends on certain assumptions when going through 'subdirs' target
template. Each subdir needs to have a project file with the name
"$SUBDIR/$SUBDIR.pro" - and if one is not found, qmake will skip the
directory silently.

After syncing the submodules into the checked out tree, see if each
subproject has a matching .pro file. If one is not found, create a
symbolic link with the expected name that points to the actually
discovered .pro file.

This satisfies qmake's assumption, and allows to properly recurse into
each of the subprojects.